third_party/WebKit/LayoutTests/http/tests/security/promise-access-control-deny-expected.txt

CONSOLE ERROR: line 1: Uncaught (in promise) 42
Tests that promise rejection events are muted for cross origin non-CORS scripts
